tests/wpt/web-platform-tests/app-history/navigate/reload-state-undefined.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    const navState = { key: "value" };
    const navInfo = { infoKey: "infoValue" };
    i.contentWindow.appHistory.navigate("#1", { state: navState }).committed.then(t.step_func(() => {
      // Make sure that state setting worked
      assert_equals(i.contentWindow.appHistory.current.getState().key, "value");
      assert_not_equals(i.contentWindow.appHistory.current.getState(), navState);

      let start_url = i.contentWindow.location.href;
      let start_key = i.contentWindow.appHistory.current.key;
      let start_id = i.contentWindow.appHistory.current.id;
      let onnavigate_called = false;
      let promise_settled = false;
      i.contentWindow.appHistory.onnavigate = t.step_func(e => {
        e.transitionWhile(Promise.resolve());
        onnavigate_called = true;
        assert_equals(e.info, navInfo);
        assert_equals(e.navigationType, "reload");
        assert_equals(e.destination.getState().key, "value", "destination.getState()");
        assert_not_equals(e.destination.getState(), navState);
      });
      i.contentWindow.appHistory.reload({ info: navInfo, state: undefined }).committed.then(t.step_func_done(() => {
        assert_true(onnavigate_called);
        assert_equals(i.contentWindow.location.href, start_url);
        assert_equals(i.contentWindow.appHistory.current.key, start_key);
        assert_not_equals(i.contentWindow.appHistory.current.id, start_id);
        assert_equals(i.contentWindow.appHistory.current.getState().key, "value", "destination.getState()");
        assert_not_equals(i.contentWindow.appHistory.current.getState(), navState);
      }));
    }));
  });
}, "reload() variant with info and state: undefined counts the same as not present (because of Web IDL dictionary semantics), so preserves the state");
</script>
